pyproject.toml
[tool.poetry]
name = "fast-api-2"
version = "0.1.0"
description = ""
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = "^2.0.24"
python-dotenv = "^1.0.0"
alembic = "^1.13.1"
psycopg = "^3.1.16"
asyncpg = "^0.29.0"
postgres = "^4.0"
pytest = "^8.0.2"
httpx = "^0.27.0"
yarl = "^1.9.4"
pytest-asyncio = "^0.23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"